+++ Ko-Fi Engine/Source/Material.cpp
@@ -1,13 +1,48 @@
 #include "Material.h"
+#include "FSDefs.h"
 #include "Color.h"
 #include "Texture.h"
+#include <glew.h>
 
-Material::Material()
+Material::Material() :
+	shaderProgramID(0),
+	shaderPath(SHADERS_DIR + std::string("default_shader") + SHADER_EXTENSION)
 {
-
 }
 
 Material::~Material()
 {
+	glDeleteProgram(shaderProgramID);
 
+	uniforms.clear();
+
+	shaderPath.clear();
+	shaderPath.shrink_to_fit();
+}
+
+Uniform* Material::FindUniform(std::string name)
+{
+	for (Uniform* uniform : uniforms)
+	{
+		if (uniform->name == name)
+		{
+			return uniform;
+		}
+	}
+	return nullptr;
+}
+
+void Material::AddUniform(Uniform* uniform)
+{
+	uniforms.push_back(uniform);
+}
+
+void Material::UseShader()
+{
+	glUseProgram(shaderProgramID);
+}
+
+void Material::DeleteShader()
+{
+	glDeleteProgram(shaderProgramID);
 }
