+++ Ko-Fi Engine/Source/Material.h
@@ -2,9 +2,25 @@
 #define __MATERIAL_H__
 
 #include <string>
+#include <vector>
+#include "Globals.h"
 #include "Color.h"
 
 class Texture;
+
+struct Uniform
+{
+	Uniform(std::string name, uint type) : name(name), type(type) {};
+	uint type;
+	std::string name;
+};
+
+template<class T>
+struct UniformT : public Uniform
+{
+	UniformT(std::string name, uint type, T value) : Uniform(name, type), value(value) {}
+	T value;
+};
 
 class Material
 {
@@ -12,13 +28,27 @@ public:
 	Material();
 	~Material();
 
-	inline void SetColor(Color color) { diffuseColor = color; }
+	Uniform* FindUniform(std::string name);
+	void AddUniform(Uniform* uniform);
+
+	void UseShader();
+	void DeleteShader();
+
+	inline const char* GetShaderPath() const { return shaderPath.c_str(); }
+	inline void SetShaderPath(const char* name) { this->shaderPath = name; }
+
+	//inline void SetColor(Color color) { diffuseColor = color; }
 
 public:
-	Color diffuseColor;
+	uint shaderProgramID;
+	std::vector<Uniform*> uniforms;
+
+	//Color diffuseColor;
 
 	std::string materialName;
 	std::string materialPath;
+
+	std::string shaderPath;
 };
 
 #endif // !__MATERIAL_H__
